backend/pyproject.toml
[tool.poetry]
name = "music-assist-backend"
version = "0.1.0"
description = "Music-Assist Backend API"
authors = ["Admin <admin@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.104.1"
uvicorn = {extras = ["standard"], version = "0.24.0"}
python-multipart = "0.0.6"
pydantic = "2.5.0"
langchain = "^0.1.0"
langchain-openai = "^0.1.0"
langchain-community = "^0.0.10"
openai = "^1.6.1"
faiss-cpu = "1.7.4"
aiohttp = "3.9.1"
beautifulsoup4 = "4.12.2"
lxml = "4.9.3"
numpy = "1.24.3"
pandas = "2.0.3"
python-dotenv = "1.0.0"
firebase-admin = "6.2.0"
python-json-logger = "2.0.7"

[tool.poetry.scripts]
start = "uvicorn main:app --reload"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"